src/test/java/com/taurinustrading/dogapi/web/DogControllerUnitTest.java
package com.taurinustrading.dogapi.web;

import com.taurinustrading.dogapi.controller.DogController;
import com.taurinustrading.dogapi.service.DogService;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.security.test.context.support.WithMockUser;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.web.servlet.MockMvc;

import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@ExtendWith(SpringExtension.class)
@WebMvcTest(DogController.class)
@WithMockUser(username = "admin", password = "password", authorities = "USER")
public class DogControllerUnitTest {
    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private DogService dogService;

    @Test
    public void getAllDogs() throws Exception {
        mockMvc.perform(get("/dogs"))
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON))
                .andExpect(content().json("[]"));

        verify(dogService, times(1)).findAllDogs();
    }

    @Test
    public void getAllDogBreeds() throws Exception {
        mockMvc.perform(get("/dogs/breeds"))
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON))
                .andExpect(content().json("[]"));

        verify(dogService, times(1)).findAllDogBreeds();
    }

    @Test
    public void getBreedById() throws Exception {
        mockMvc.perform(get("/dogs/1/breeds"))
                .andExpect(status().isOk());

        verify(dogService, times(1)).findDogBreedById(1L);
    }

    @Test
    public void getAllDogNames() throws Exception {
        mockMvc.perform(get("/dogs/names"))
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON))
                .andExpect(content().json("[]"));

        verify(dogService, times(1)).findAllDogNames();
    }
}
